Now compiles on MSVC++
git-svn-id: svn://svn.code.sf.net/p/annchienta/code/trunk@558 6c58e83f-43c0-4801-a1bd-44083660f6ae

+++ Annchienta/annchienta/src/point.cpp
@@ -158,7 +158,7 @@ namespace Annchienta
     int Point::distance( Point other ) const
     {
         Point tmp = other.to( type );
-        return (int) sqrt( square(x-tmp.x) + square(y-tmp.y) + square(z-tmp.z) );
+        return (int) sqrt( double(square(x-tmp.x) + square(y-tmp.y) + square(z-tmp.z)) );
     }
 
 };
